requirements.txt
ultralytics
pillow
PyQt6
